backend-FASTAPI/pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosmtplib>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.121.0",
    "mistralai>=1.9.11",
    "pydantic>=2.12.4",
    "pymongo[srv]>=4.15.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
